--- frontend/src/components/TextField/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './TextField';
