+++ frontend/src/components/layout/Menu/Menu.tsx
@@ -4,11 +4,20 @@ import { ListItemIcon, ListItemText, ListItemButton } from '@mui/material';
 
 import { MenuTabs } from './styled';
 
-import { menuItems } from './MenuContainer';
+interface MenuItemProps {
+  path: string;
+  title: string;
+  icon: React.ReactElement;
+}
 
-const Menu: React.FC = () => (
+interface MenuProps {
+  menuList: MenuItemProps[];
+  children?: React.ReactNode;
+}
+
+const Menu: React.FC<MenuProps> = ({ menuList }) => (
   <MenuTabs>
-    {menuItems.map((item, key) => (
+    {menuList.map((item, key) => (
       <ListItemButton key={key} component={Link} to={item.path}>
         <ListItemIcon>{item.icon}</ListItemIcon>
         <ListItemText primary={item.title} primaryTypographyProps={{ variant: 'h6' }} />
