+++ frontend/src/components/layout/Menu/MenuContainer.tsx
@@ -1,100 +1,13 @@
 import React from 'react';
-import {
-  AccountCircleOutlined,
-  ArticleOutlined,
-  FolderSpecialOutlined,
-  AddToQueueOutlined,
-  EmojiPeopleOutlined,
-  AssessmentOutlined,
-  QuizOutlined,
-} from '@mui/icons-material';
 
-import PATHS from 'constants/routes';
+import { ROLES_MENU } from 'constants/menuRoles';
 
 import Menu from './Menu';
 
-const EMPLOYEE_MENU = [
-  {
-    path: PATHS.profile,
-    title: 'Profile',
-    icon: <AccountCircleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.coursesList,
-    title: 'Courses List',
-    icon: <ArticleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.myCourses,
-    title: 'My Courses',
-    icon: <AddToQueueOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.help,
-    title: 'Help',
-    icon: <QuizOutlined fontSize="large" />,
-  },
-];
+const MenuContainer: React.FC = () => {
+  const menuItems = ROLES_MENU['employee'];
 
-const MANAGER_MENU = [
-  {
-    path: PATHS.profile,
-    title: 'Profile',
-    icon: <AccountCircleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.coursesList,
-    title: 'Courses List',
-    icon: <ArticleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.myCourses,
-    title: 'My Courses',
-    icon: <FolderSpecialOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.requests,
-    title: 'Pending Requests',
-    icon: <AddToQueueOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.employees,
-    title: 'Employees',
-    icon: <EmojiPeopleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.help,
-    title: 'Help',
-    icon: <QuizOutlined fontSize="large" />,
-  },
-];
-
-const ADMIN_MENU = [
-  {
-    path: PATHS.coursesList,
-    title: 'Courses List',
-    icon: <ArticleOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.skills,
-    title: 'Skills',
-    icon: <AssessmentOutlined fontSize="large" />,
-  },
-  {
-    path: PATHS.help,
-    title: 'Help',
-    icon: <QuizOutlined fontSize="large" />,
-  },
-];
-
-const ROLES_MENU = {
-  employee: EMPLOYEE_MENU,
-  manager: MANAGER_MENU,
-  admin: ADMIN_MENU,
+  return <Menu menuList={menuItems} />;
 };
-
-export const menuItems = ROLES_MENU['employee'];
-
-const MenuContainer: React.FC = () => <Menu />;
 
 export default MenuContainer;
